+++ domain/src/main/java/dev/jameido/pokedex/domain/entity/PkmnDetailEntity.kt
