+++ domain/src/main/java/dev/jameido/pokedex/domain/entity/PkmnListEntity.kt
