+++ domain/src/main/java/dev/jameido/pokedex/domain/entity/PkmnVarietyEntity.kt
