+++ settings.gradle
@@ -1,2 +1,3 @@
+include ':domain'
 include ':app'
 rootProject.name = "Pokédex"
